fix: run gh pr view instead of create for revisions with known PRs
